BlogProject/requirements.txt
dj-database-url==0.5.0
Django==2.1.5
django-ckeditor==5.6.1
django-heroku==0.3.1
django-js-asset==1.1.0
gunicorn==19.9.0
Pillow==5.4.1
pkg-resources==0.0.0
psycopg2==2.7.7
pytz==2018.9
whitenoise==4.1.2
